requirements.txt
python-telegram-bot==13.8.1
google-cloud-dialogflow==2.9.1
python-dotenv==0.19.1
vk-api==11.9.6